requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
scipy>=1.6.0
pandas>=1.2.0
vidgear>=0.1.9
progressbar2>=3.53.0 